requirements.txt
pytesseract~=0.3.10
Pillow~=10.0.0
email-validator==2.0.0.post2
